DeleteWorkflowExecution is not transactional with workflow update
When decider responds back with complete workflow decision, we first
update the execution with new events and then delete workflow execution
as a separate transaction.  This can cause issues when the update times
out but we successfully apply the update.  This can cause us to never
delete this workflow execution.
History engine now creates a transfer task for processing the delete in
the same transaction as update.  Transfer queue processor is updated to
process these delete execution tasks and actually delete the workflow
execution from store.
I have also cleaned up matching engine to not use ActivityTask and
DecisionTask which are being used for transfer task processing.