Merge GetWorkflowExecution and GetWorkflowMutableState
We have moved Workflow Execution history to a separate table now and is
no longer persisted as a blob in executions table.  Now all the workflow
execution state in executions table represents the mutable state for the
workflow execution.  We no longer need to separate APIs to read
ExecutionInfo and MutableState.