Frontend support for domain resolution from name to uuid
Added DomainCache as a common component which can be used by
frontend for resolving domain name to domain uuid, which is used
by history and matching engines for identifying the domain and
making persistence API calls.
Frontend is not passed in MetadataManager, which is used for
implementing CRUD for managing domains.
Interation tests now also creates a domain as part of test setup.