Engine changes to support signal workflow execution
Engine implementation for writing the workflow execution signaled event
to history.  Refactored code to reuse the update workflow logic between
terminate and signal implementations.
Persistence layer changes to retrieve the current run_id for an
execution.  Update history cache to first call the GetCurrentExecution
persistence API for operations like Terminate and Signal when run_id is
not specified as part of the request.