Persistence layer changes to support ContinueAsNew
UpdateWorkflowExecution now optionally takes in
StartWorkflowExecutionRequest as ContinueAsNew parameters.
CreateWorkflowExecution is also refactored to move out the query logic
into another function so both Create and Update can reuse it.  If
CreateWorkflowExecution is called with ContinueAsNew it updates the
current_run_id for the permanentRunID row for the execution.
Fixed StartWorkflow integration test to handle multiple starts with the
same request to not return an error.