TransferQueueProcessor can skip transfer tasks
TransferQueueProcessor uses Quorum level consistency to read tasks.
These reads are not isolated from in-flight LWT transactions running
in Cassandra, which could lead the read to observe partial results.
This can lead to tasks being skipped.

We can use Serial level consistency to deal with this but it is not
supported at the moment by the gocql client, and it might cause reads
to be too expensive. Instead, we'll use a MaxReadLevel to make sure
that the read requests from TransferQueueProcessor do not attempt to
read the results of LWT writes that are still in flight.