Matching engine fix for flaky unit test
TaskListManager needs range to updated before starting task writer so it
can initialize max read level correctly using the initialized
taskSequenceNumber on TasklistMgr.  If ordering of reversed then
MaxReadLevel is initialized incorrectly to -1 and will rely on a task to
be written before it can be set to correct value.  This results in
GetTasks to use incorrectly initialized MaxReadLevel and will never read
tasks from persistence store.