TimerQueueProcessor stuck fix on large backlog on available timers
TimerQueueProcessor issues a query to DB for visibilityTime '>=' from
the readlevel.  This could result in reading same timer multiple times
and trips the processor in thinking it fired all available timers as the
number of tasks returned back to the caller would be less than batch
size.
Updated the logic to return an explicit flag back to the caller so it
can immediately callback again when there are more timers to fire.