Support for generating replication task on workflow execution updates
History engine changes to generate replication task when IsGlobalDomain
is enabled.  MutableStateBuilder is also updated to load replication
state for the domain and have support for creating replication task on
updates along with relevant updates to mutable state.

Replication tasks are written to a separate queue for the shard.  This
requires persistence changes to write replication tasks on updates in
the same transaction and new API for reading and completing the task.
New AckLevel is needed for processing of replication tasks.  This
requires changes to shard crud on the persistence layer for reading and
writing the new ack level for replication task processing.

ClusterMetadata and MessagingClient are already exposed on the Service.
So passing through Service to all layers in History service which needs
access to both ClusterMetadata and MessagingClient.