Use correct timestamp on transient decision history events
We store the timestamp of when decision is started on transient
decision within mutable state.  But when generating the history event we
don't use it at all.  This could result in non-deterministic error on
the client as it relies on timestamp on the decision task started event.
This fix uses the timestamp stored on mutablestate for transient
decisions.