Expose ResetStickyTaskList API in frontend
This commit exposes the ResetStickyTaskList API, which up until now
has been internal only and resides in history service, in the frontend service.

The code that exposes this API in frontend is almost 100% mechanical and acts
as practically as a forwarder.

The motivation for exposing this previously internal API is to clear the way
for future client side changes that would involve client side workers explicit
clearing the stickiness of their workflows, for example, due to potentially newly
introduced logic that evicts workflow contexts based on a timer.