Add ActivityScheduleTimeout deduction logic
This commit relaxes validateActivityScheduleAttributes by having it fill in some
of the fields related to schedule timeouts if they are empty.

The logic used is identical to the logic used in AWS SWF's java client.

The logic is as follows:
If ScheduleToClose is present and ScheduleToStart or StartToClose is missing,
we will set what is missing to ScheduleToClose.

If ScheduleToClose is missing but both ScheduleToStart and StartToClose are present,
we will set ScheduleToClose to the sum of ScheduleToStart and StartToClose.