Return ShardOwnershipLostError as catch all from persistence layer
Cadence history service rely on ShardOwnershipLostError for triggering
unload of the shard.  If persistence layer is unable to detect
ShardOwnershipLost error from Cassandra then it will cause retries
within Timer and Transfer processors and could result in fatal error
causing the entire host to be restarted.
Updating the persistence layer to always return ShardOwnershipLostError
is a catch all when unable to detect the cause of failed transaction to
only cause reload of single shard.