Replacing glide with dep
Dep generates a folder, _vendor-*, with copies of the source code for
dependencies. Added this to .gitignore and set the "make copyright"
check to skip that folder.

Not sure if it's related but "make bins" fails to build thrift because
there is no Go file in the thriftrw folder. Changed the yarpc build to
just build thrift/wire.