Reorder activity completion events to process at the end
Sometimes we see buffered events could be out of order when loaded from
database.  This is not an issue on most cases but becomes a problem when
activity completion events is processed before activity started event.
This change reorders the processing of buffered events to always process
activity completion events at the end.