Move history events out of mutableState to eventsCache
Storing history events within mutableState limits scalability of the
system.  So moving all history events out of mutableState to
separate eventsCache.  This makes the mutable state size to be quite
small and allows us to store very large number of activities.
EventsCache is a LRU cache which stores the history event.  Any
processing logic which needs the history event like
RecordActivityStarted can get it from the cache without reading
it from the database.  Incase event is not found in cache it
uses the HistoryManager to load it from database and return it
to caller.