Allow query task for non-active domains
Cadence matching completely stops dispatching of both DecisionTask
and QueryTask in response to PollForDecisionTask API when domain
is marked as non-active.
We have separated out the task dispatch channel for query and
decisions, as we always want to allow sync match on queries even
if the domain is not active.