Archiver GCloud Storage
We have developed a gcloud archiver provider in order to push historical workflows files to Google cloud storage.
This feature will allow us to move on all history to a single place, very useful when you are running cadence
as Cluster mode. In order to accomplish this task we have developed the following modules:
1. archiver/provider -> based on a deployment configuration file will instanciate a gcloud historyArchiver
2. archiver/gcloud/historyArchiver -> implements archiver.HistoryArchiver interface
3. archiver/gcloud/connector -> gcloud storage client implementation.