Increment readLevel when skipping over expired tasks
TaskReader has to logic to skip over expired tasks instead of
adding it to AckManager to prevent dispatch.  This results in
taskReader to get stuck in all tasks read from the database are
expired as it results in readLevel to not move at all.
This change also increments the readLevel when skipping expired
tasks to prevent taskReader to spin on expired tasks.