Add header options for starting workflow via CLI
It was not possible to set header values via CLI when starting a workflow.

This change exposes three new optional parameters for starting a workflow:
- header_key
- header_value
- header_file

They allow setting header values in JSON format which are propogated
via workflow context. Values can be provides either directly or with a
file. If multiple keys and values are used, they separated by a space.
Count of keys and values must be equal, otherwise error is returned.