k8s: fix cassandra-tool env key conflict with k8s
1.the default CASSANDRA_PORT key is used by k8s when you have an
service pod named cassandra(k8s will gen a env key which is name_port).
And the CASSANDRA_PORT has valid value which is an url, cannot be i
parsed to an int value. Finally, the cadence-cassandra-tool can not be
executed. Candence cannot work well when need cadence-canssandra.
2.here I just raplace the CASSANDRA_PORT to CASSANDRA_DB_PORT, to avoid
use the same key with k8s. Or is there any key more fit?

Depends-On: N/A