k8s: fix cassandra-tool env key conflict with k8s (#3505)
*  k8s: fix cassandra-tool env key conflict with k8s

1. The default CASSANDRA_PORT key is used by k8s when you have an
service pod named cassandra(k8s will gen a env key which is name_port).
And the CASSANDRA_PORT has valid value which is an url, cannot be 
parsed to an int value. Finally, the cadence-cassandra-tool can not be
executed. Candence cannot work well when need cadence-canssandra.
2. Replace the CASSANDRA_PORT to CASSANDRA_DB_PORT, to avoid
use the same key with k8s. 

* tools: terminated with a non-zero exit code

1. When cannot run
2. These tools depend on some environment vars, which are not really
reliable. In that case, tools start and stop without any stdout.