Simplify tags
Just noticed while skimming that this file seemed drastically over-complicated.
As the keys are all consts, this should mean that the net change for all this is
that these take one slice pointer more space per instance.

Separately, now that it's simpler, some of the variations are more noticeable.
It may be worth standardizing further, e.g.:
- sanitizing everything (except perhaps the kafka partition)
- get rid of the "unknown" constructors?  I can see them having some use, but it seems kinda random what has it and what doesn't.