Ignore comment-formatting lints
This does not cover all of the comment-related lints, but it does cover all of the ones that currently produce warnings.
After this, 10 go away, bringing the total to 18.

**Rationale:**

Many of the comment-requiring and comment-formatting lints are focused on both:
1. Encouraging documenting public APIs.  But this repo is effectively entirely private / not intended for people to import.
2. Producing "nice" documentation in `go doc` output.  Whether this makes more readable docs or not is debatable, but I understand the goal.

When *either or both* of those don't apply, the lints just kinda suck.  Their mere existence encourages pointless stuff like:
```go
// ConvertPretty ...
<snip>
func (e *ESql) ConvertPretty(sql string, pagination ...interface{}) (dsl string, sortField []string, err error) {
```
Or this, which appears in 30 places:
```go
// Ptr is a helper function for getting pointer value
func ... Ptr() ...
```
Or this TBD pattern, in **OVER 2000** places:
```go
// WorkflowExecutionCanceledEventAttributes is an internal type (TBD...)
```

It's a waste of time, space, and mental-energy.

In my experience it also implies the code is in a "higher quality" state than it may actually be, as good code often has documentation.
While that's obviously not *actually true*, I've frequently seen people stop reading when they find a documented func regardless of what the documentation actually says, on the assumption that it's either 1) functioning correctly, or 2) intended for broader use.
2 in particular often manifests as code importing some random unrelated package because it happens to have the only documented + commonly-used pointer-conversion-func.  I've seen this in dozens of repositories, it's a plague.

Even if we (think we) aren't vulnerable to that kind of thinking when writing code, it seems worth avoiding since it clearly brings no benefit.

Document what needs documentation.
Don't add documentation that offers nothing beyond the type name (or less).
And let's delete bad docs when we see them.