Use a test-logger in tests rather than stdout (#3976)
Currently, on failure, e.g. matchingEngine_test.go tests spew out *ridiculous* amounts of text.
This happens because Go cannot tell which logs are related to the failed test(s), so it does what it can: show all stdout.
Unfortunately, stdout blends logs from parallel tests, and it's hard or impossible to find the ones relevant to the failure simply by reading.
So: stop doing that.  Zap has a `zaptest` package for this purpose.
I'm not sure how many other places may need to be changed, but this helps the ones I'm looking at at the moment.