Move tcheck to go.mod, get rid of glide dependency (#3999)
[glide.sh is down](https://github.com/Masterminds/glide/issues/1077), breaking our pipelines, and glide is abandonware anyway.
It's well past time to migrate off.

I don't believe `tcheck` is actually used for anything, nor is it mentioned in any documentation... but it _does_ work, it'll tell you if e.g. cadence-frontend is running and could be useful for scripting.  Though in that case a relevant `cadence ...` command would be as good or better anyway.

It may be worth removing this entirely?  It's not particularly large nor pulling in any new dependencies, but it may just be old cruft.