Add scavenger for orphaned tasks
TaskList entries can be deleted out of the task_lists table before the corresponding tasks are drained from the tasks table. This leaves those tasks orphaned, and prior to this change nothing will delete them so they accumulate.
This change enhances the tasklist scanner to watch for these orphans and remove them.