Bugfix: write changes back to *Persistence.DataStores
Since DataStores is a `map[string]DataStore`, `ds` is a copy of the data in the map, and changes are made only to that copy.
So this change writes the (possibly)-changed value back to the map, so it actually takes effect.

I added a test because why not.  Prior to this fix it fails with:
```
--- FAIL: TestCassandraFallbackConfig (0.00s)
    config_test.go:90:
                Error Trace:    config_test.go:90
                Error:          Should NOT be empty, but was <nil>
                Test:           TestCassandraFallbackConfig
                Messages:       nosql config should contain cassandra config
```
and now it works.