Bugfix: write changes back to *Persistence.DataStores (#4284)
Since DataStores is a `map[string]DataStore`, `ds` is a copy of the data in the map, and changes are made only to that copy.
So this change writes the (possibly)-changed value back to the map, so it actually takes effect.

I added a test because why not.  Prior to this fix it fails with:
```
--- FAIL: TestConfigFallbacks (0.00s)
    config_test.go:95: 
                Error Trace:    config_test.go:95
                Error:          Should NOT be empty, but was <nil>
                Test:           TestConfigFallbacks
                Messages:       nosql config should contain cassandra config / not be empty
```
and now it passes.

The NumShards update worked before because it's modifying a `*SQL`, but I've included an assertion for it to serve as a counter-example / to guard against a bug if it ever becomes a value object as well.  Plus I assume we have other "fill in defaults / backfill legacy stuff" values, or we will eventually, and bundling all the simple ones into one test is fairly reasonable.