Show more information about panicked objects
Well that's kinda a ridiculous thing to do.
Oh well, it's fixed now.

While checking other uses of `recover()`, I also found one highly
suspicious piece of test code that was suppressing panics and allowing
the test to run successfully.  I have not read the surrounding code
further, but this also seems ridiculous at a glance, so it's gone too.