Improve error message when a workflow's decision is rejected due to concurrent changes (#4673)
When a workflow e.g. receives a signal while processing a decision task, the task's result
will not be applied, and an 'Unhandled decision' error will be returned.  The next decision
task will (hopefully) produce a new set of decisions that take that signal into account.

When this error is returned, in the UI it currently shows as a decision-task-failed with
these details:
```
cause   UNHANDLED_DECISION
details []
```
which is extremely unclear to users.  After this change, hopefully it will be easier to
understand.