Changes error handling for ServiceIsBusyError to use go 1.13 error conventions
As per https://go.dev/blog/go1.13-errors using error unwrapping with
Is/As allows for richer information to be embedded in returned errors,
and would (in the future) allow developers to add contextual information
without loss of fidelity in control flow.

This PR is 1/n for this particular error. It's intended as a spike to
determine how difficult such a change is to make, and a path for future
error handling.