As-simple-as-possible addition of background-processing ratelimits
"User" requests have been ratelimited for basically forever, but requests
from workers (e.g. polling for tasks, responding) have not, except at the
most coarse level (per-host RPS).

This has left us with no effective way to clamp down on heavy-weight domains
when systems are becoming overloaded.  While we can limit how fast they
*begin* new workflows, they continue *processing* as quickly as possible,
particularly if there is a backlog (e.g. they started several million
workflows at once to work through a bulk job of some kind).

Since we are in need of a quick "stop the bleeding" patch, this is built
roughly as simply as I could manage.  All apparently-relevant ratelimits
now split "user" and "worker" requests apart, and throttle them individually.
The defaults for "worker" requests are `0` (unlimited) so this should be a
no-op until configured to do otherwise.

This also moves many "counted but not limited" endpoints to the per-domain
"worker" limit, so they can be controlled.  Since the default is unlimited,
and the per-domain limiter also increments the global counts, this should
also be a no-op but makes things more flexible in configuration.

To both simplify identifying things that need changing, and to make the
user/worker separation more clear, I've renamed all relevant "RPS" to
"UserRPS" and added "WorkerRPS" (and similar renames).

I am not confident that this covers all possible endpoints