Reformatting most things for go 1.19, rebuilding go.mod tools after clean, warning about different go versions (#5019)
The .bin folder isn't cleaned and doesn't version itself based on all possible influences...
... so it's kinda just turning into too much of a headache when it misbehaves.

In particular, Go versions sometimes change formatting, and that tends to be baked into
formatting tools, rather than executing `go` externally.  So when e.g. `goimports` is built
with a different version of Go than you are currently using, things can get confusing.
These are now in the .build folder, so `make clean` will force a rebuild.

Downloaded bins like `buf` and `protoc` are pre-built and should generally be resistant to
this kind of leakage... and forcing a re-download on every `make clean` seems like definitely
a bad idea in some scenarios.  So the .bin folder remains, but its use is now much smaller.

---

For gofmt purposes, formatting with the latest go version _tends_ to be stable...
but unfortunately 1.19 and 1.17 disagree on protoc's code block indentation in comments.

As CI currently runs 1.17, we need to be stable with 1.17's output.
So this commit does a 1.19 format, then a 1.17 format.  This appears stable with our
hand-crafted comments, and is likely an improvement over just 1.17.

It does, unfortunately, mean that you truly must use <1.19 or your commits may not be stable.
We will likely be upgrading to 1.19 soon, but that involves a number of changes, so it's not
done here.  Stability first, upgrades second.

---

The warning is printed immediately for all commands, but does not stop you from using that version:
```
❯ make
Makefile:41: Caution: you are not using CI's go version. Expected: go1.17, current: go version go1.18 darwin/amd64
help                 Prints a help message showing any specially-commented targets
...
```