Support dynamic header-forwarding rules, add list-based dynamic config
Blindly copying all inbound headers to outbound is unamibguously wrong, but the exact list that //is// appropriate is complex and not necessarily the same everywhere.
So this adds a semi-flexible config-driven rule for controlling this.

It is in dynamic config currently because it's easier for us to change internally, while trying to build a smarter list of headers - it belongs in static config, and possibly dynamic config *after refreshing is added*.