Adds isolation-group and partition libraries (#5262)
Adds two libraries for the 'Zonal-isolation' feature:

Isolationgroup
This is a library with two sections: state which is to be used throughout the Cadence Frontend and Matching applications to determine what the current state of isolation-groups is. it's intended to be stateful daemon, though this first implementation doesn't directly contain state, but it is intended to be extended in that matter as a singleton which uses domain-cache and config-store below it to persist data. The other section is api which is a simple CRUD interface for interacting with the two sources of data.

Partition
This is a library and some default logic for dividing up task processing in a reasonably sensible default:

The isolation-group in which the workflow is created is where it will remain, unless there's a drain.

This also introduces a simple distribution mechanism for spreading processing deterministically over the remaining isolation-groups should the original one indeed be drained.