Validate search-attribute-key so keys are fine in advanced search
ElasticSearch engine fails to parse keys with dashes, making those
search-attributes broken for CLI and web.
Validate search-attribute before creating on server so it looks like a
valid variable name in most of languages, should be more than expected for
programmers.

The reason it is not in the server, is that currently even at Uber we
have keys with dashes, and it will be a problem if we gonna migrate
domains between clusters.